:truck: Rename : logger 폴더로 캡슐화
